Move calculate to derange.ts
