+++ src/derange.ts
@@ -68,3 +68,16 @@ export const derange = (
     return buffer2[personIndex];
   });
 };
+
+export const calculate = async (
+  people: Person[],
+  exclusions: Exclusion[] = []
+): Promise<Person[]> => {
+  return new Promise((resolve, reject) => {
+    try {
+      resolve(derange(people, exclusions));
+    } catch (e) {
+      reject(e);
+    }
+  });
+};
