+++ test/derange.test.ts
@@ -1,7 +1,7 @@
 import '../to-be-valid-derangement';
-import { derange, validateMatches } from '../src/derange';
+import { calculate, derange, validateMatches } from '../src/derange';
 import { personArrayOfLength } from '../src/utils';
-import { Exclusion, Person } from '../src/models';
+import { DerangementError, Exclusion, Person } from '../src/models';
 
 describe('derange', () => {
   describe('validateMatches', () => {
@@ -291,6 +291,36 @@ describe('derange', () => {
       ];
 
       expect(derange(input, exclusions)).toBeValidDerangement(input);
+    });
+  });
+
+  describe('calculate', () => {
+    it('calculates exchanges', async () => {
+      const input = personArrayOfLength(50);
+      const result = await calculate(input);
+      expect(result).toBeValidDerangement(input);
+    });
+
+    it('rejects Promise when given impossible exclusions', async () => {
+      const input = personArrayOfLength(3);
+      const exclusions: Exclusion[] = [
+        {
+          type: 'name',
+          subject: '2',
+          value: '1'
+        },
+        {
+          type: 'name',
+          subject: '1',
+          value: '2'
+        }
+      ];
+      try {
+        await calculate(input, exclusions);
+        expect(true).toBeFalsy();
+      } catch (e) {
+        expect(e instanceof DerangementError).toBeTruthy();
+      }
     });
   });
 });
